Chiavari Precipitation/Rainfall Dataset

This dataset contains time series of obsevations of precipitation intensity with hourly timestep.

data(chiavari)

Details

The 'chiavari' variables is a data frame containg time,variable and value. Further details can be found on the URLs below.

Format

Data frames and vectors

Source

Original data are provided by Comune di Chiavari, Osservatorio del Seminario vescovile di Chiavari, in the framowork of the project "Recupero e digitalizzazione della serie storica dei dati meteo rilevati presso l'osservatorio meteorologico sismico del Seminario vescovile di Chiavari negli anni 1883 - 2014" by "Associazione Amici del Museo Scientifico G. Sanguineti - G. Leonardini" , https://servizi.comune.chiavari.ge.it/osservatorio/ for the period 1883-2014 and ARPA (Environmental Protection Regional Agency) Regione Liguria, https://ambientepub.regione.liguria.it/SiraQualMeteo/script/PubAccessoDatiMeteo.asp directed from http://www.banchedati.ambienteinliguria.it/index.php/aria/meteo fort the periond 2014-2022.

This dataset is intended for research purposes only, being distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY.